types: Make QdlChan BufRead
The Read trait serves us well for the USB backend, but the serial backend
doesn't provide incoming messages in packets. This becomes a problem when we're
doing <read/> as the response and data can come in the same channel.read().

Switch the trait to BufRead, to allow the implementation to peek at the data and
partially consume the channel.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>