qdl: Peek at the BufRead in firehose_read()
The channel does not necessarily provide the incoming messages in the form of
packetized messages. This means that the response XML and rawmode data might be
read at once. As we switched to a BufRead we can look for the </data> marker and
consume only that amount of data instead.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>