qdl: Handle non-chunk-sized reads
Relying on reads always returning sector-sized chunks, doesn't work when the
underlying transport in qcserial. In the upcoming changes, the reader is changed
to BufRead which exposes this problem.

There's nothing in the loop that rely on the reads always returning sector-sized
chunks, so drop that part of the logic.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>